+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## Not released
 
+- Fix widget calculation with very large viewports/masks [#680](https://github.com/CartoDB/carto-react/pull/680)
+
 ## 2.0
 
-## 2.0.4 (2023-05-19)
+### 2.0.4 (2023-05-19)
 
 - Fix type propTypes issues [#677](https://github.com/CartoDB/carto-react/pull/677)